+++ changelog.md
@@ -1,6 +1,9 @@
-# Release Notes
+﻿# Release Notes
 
 ## SmartStore.NET 2.5.1
+
+### Highlights
+* #637 Integrate PayPal PLUS Provider
 
 ### New Features
 * #571 Option to display another checkbox on confirm page to let the customer accept that his email address can be handed over to a third party